convert lessons paths to /module
