+++ apps/epic-react/src/pages/modules/[module]/[lesson]/exercise.tsx
