+++ apps/epic-react/src/pages/modules/[module]/[lesson]/solution/index.tsx
