+++ apps/epic-react/src/pages/modules/[module]/index.tsx
